refactor: move xcm config to separate file
